requirements.txt
fastapi
uvicorn[standard]
pydantic
openai
python-dotenv
streamlit
requests